+++ com.unity.render-pipelines.universal/Runtime/ForwardRenderer.cs
@@ -163,8 +163,8 @@ namespace UnityEngine.Rendering.Universal
                 generateColorGradingLUT = false;
 #endif
 
-            bool isSceneViewCamera = cameraData.cameraType == CameraType.SceneView;
-            bool isPreviewCamera = cameraData.cameraType == CameraType.Preview;
+            bool isSceneViewCamera = cameraData.isSceneViewCamera;
+            bool isPreviewCamera = cameraData.isPreviewCamera;
             bool requiresDepthTexture = cameraData.requiresDepthTexture;
             bool isStereoEnabled = cameraData.isStereoEnabled;
 
@@ -535,7 +535,7 @@ namespace UnityEngine.Rendering.Universal
             if (cameraData.renderType == CameraRenderType.Base && !cameraData.resolveFinalTarget)
                 return true;
 
-            bool isSceneViewCamera = cameraData.cameraType == CameraType.SceneView;
+            bool isSceneViewCamera = cameraData.isSceneViewCamera;
             var cameraTargetDescriptor = cameraData.cameraTargetDescriptor;
             int msaaSamples = cameraTargetDescriptor.msaaSamples;
             bool isStereoEnabled = cameraData.isStereoEnabled;